+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.1][] - 2020-10-03
+
+Version 0.7.1 fixed an issue with working copy renames on old versions of Git.
+
+[0.7.1]: https://github.com/gg-scm/gg-git/releases/tag/v0.7.1
+
+### Fixed
+
+-  Git versions before 2.18 have a bug where they will report renames even if
+   `--disable-renames` is passed. `*Git.Status` will rewrite these to an add and
+   a delete. ([#3](https://github.com/gg-scm/gg-git/issues/3))
+
 ## [0.7.0][] - 2020-10-02
 
 Version 0.7 made improvements to fetching commit information.